requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
Pillow==10.1.0
